requirements.txt
scrapy>=2.11.0
whoosh>=2.7.4
openai>=0.27.8
python-dotenv>=0.21.0 